Moved existing RmcAppBar composable to CommonUI
